adk31/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
asyncpg==0.31.0
attrs==25.4.0
Authlib==1.6.7
babel==2.18.0
backoff==2.2.1
banks==2.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
brotli==1.2.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
crewai==1.6.1
crewai-tools==1.6.1
cryptography==46.0.4
dataclasses-json==0.6.7
ddgs==9.10.0
debugpy==1.8.20
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
diffusers==0.36.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.1
fake-useragent==2.2.0
fastapi==0.123.10
fastjsonschema==2.21.2
fastuuid==0.14.0
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
google-adk==1.22.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-cloud-aiplatform==1.136.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.62.0
google-cloud-speech==2.36.1
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.62.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.14.5
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
json5==0.13.0
json_repair==0.25.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kubernetes==35.0.0
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
lancedb==0.29.2
langchain-core==1.2.9
langgraph==0.4.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.7.1
lark==1.3.1
litellm==1.80.16
llama-index-core==0.14.13
llama-index-embeddings-google-genai==0.3.2
llama-index-instrumentation==0.4.2
llama-index-readers-file==0.5.6
llama-index-workflows==2.14.1
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
notebook_shim==0.2.4
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.1
openai==2.18.0
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
platformdirs==4.5.1
portalocker==2.7.0
posthog==5.4.0
primp==0.15.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.26.7
pyOpenSSL==25.3.0
pyparsing==3.3.2
pypdf==6.7.0
pypdfium2==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytube==15.0.0
pytz==2025.2
pywin32==311
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
Send2Trash==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.50.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.4.0
tomli_w==1.2.0
toolbox-adk==0.5.8
toolbox-core==0.5.8
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uv==0.10.0
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.4
zipp==3.23.0
zstandard==0.25.0

langchain
langchain-classic
langchain-community
# langchain-core==1.2.6
langchain-ollama
langchain-tavily
langchain-text-splitters
# langgraph==1.0.9
# langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres
langgraph-checkpoint-redis
langgraph-checkpoint-sqlite
# langgraph-prebuilt==1.0.8
# langgraph-sdk==0.3.1




# langchain-core==1.2.9
# langgraph==0.4.7
# langgraph-checkpoint==4.0.0
# langgraph-prebuilt==1.0.7
# langgraph-sdk==0.3.4